Template changes + css changes + small change to default.php
Default.php:
WP-admin view for the plugin settings, changed that people see the file name to simply "default" or "minimal" when you switch between the two styles.

Added specific classes and a div to add the ability to get a border around the Compensation amount. (Which doesn't work yet currently)

CSS: Did a lot of changes to seperate Default and Minimal design and get the Minimal design to look more like the actual agreed upon design.