Radio Button checked saved + Minimal Button Design mostly done
Changed default.php to retrieve the status of Default/Minimal design to assure at least one radio button is checked when the Plugin page is reloaded.

Changed the CSS to get the Minimal Button Design done. It might require a few minor tweaks in the future, but for now it looks good.

Added an extra layer to " fake" a border to the Template Minimal. Just like with the original design.

Attempted to make a suitable switch in javascript/jQuery to get the right logo for the minimal design. Didn't work out yet. So this should be fixed.