Fix/opt out fix (#17)
* WIP pt-out fix (stuk)

* Added optin check

Changed the optin in the Templates to correspond with the data inside the HelperComponent. And added check inside the HelperComponent

* Fixes state when default = OFF (probably)

* fallback session percentage,

added percentage fallback for session percentage, added some debugging for opt-out surcharge

* Removed optin surcharge from Helpercomponent and put in co2ok plugin

Optin now works, had to migrate it from the HelperComponent to the Plugin to make it work.
This is due to having to be attached to the woocommerce hook inside the Plugin.
Also removed any debugging and unnecessary variables.
I've also added a bit of documentation

* Changed name of "optin" variables to "optout"

Changed optin to optout

* changed some public vars to private

Changed some public vars back to private. And made sure optout charge uses the same variable as the normal charge.

* Removed session percentage

Session percentage isn't set anymore in CalculatedSurcharge

* Fixes last bug, cleans up code a bit