Touch target Infohover larger, scroll down to infohover when touching (#19)
* Touch target Infohover larger, scroll down to infohover when touching

Added an "id" for the info hitarea. with a circle in css for the touch area.
Added an "id" to scroll to the infobox when you touch the info hitarea.

* Increased toucharea size and migrated scrollintoview to showInfoBox function

CSS changes to make the larger touch-area work and migrating the scrollIntoView from mouseover to showInfoBox function.

* removed media (small), added hitarea for minimal design

removed unnecessary media small for img.
added hitarea for the minimal button design as well (had to have custom values though)

* unused values removed

Removed unused values

* double var element & scrollIntoView removed

the second call of elmnt and scrollIntoView are removed.