Attempt at setting a $_GET for disable_co2ok_button_on_cart
Introducing a lot of debugging to try and find how to make the new FEAT "Disable button Cart Page Option" work.

Currently this results in "PHP: action is true" in the console and nothing succesfully ends up in the $_GET and $_POST array.

The idea is that by inserting into the browse bar you change the setting.
For example, I have:
http://localhost:8080/?page_id=14
and add:
&disable_co2ok_button_on_cart=true
At the end, this should change the setting, also with false.
Resulting in:
http://localhost:8080/?page_id=14&disable_co2ok_button_on_cart=true

This should remove the checkbox/button, but it currently doesn't.