Accessibility for screenreader and keyboard users
Added some a and button tags in template tags so you can "tab" through the co2ok button elements, it currently works entirely for the Default button, the minimal button will probably require some extra work in the CSS. (so this is currently still broken) - I've commented out the sections which I want to implement but aren't working as desired.
The chrome screen-reader can now read each relevant section for the Default button.