The Button placement on Checkout page works
The user can now select between different area's on the Checkout-page where to place the button. This selection is done on the Plugin settings page.
It works - BUT - it's still buggy.
The placement doesn't seem to work properly everywhere (but in most places it does).
The button is placed over the "Place Order" button when selecting the review_order_after_submit placement, for example.
There is one issue where the button isn't shown anywhere (it might be that this hook is removed from Woocommerce altogether)

Also, currently when using the Placement selection and saving, it won't use the other options (Default, and Shortcode) anymore.
Warning: DON'T merge the current branch until this ^ bug is fixed.

Lastly, I've put in some debugging I will remove later.