Removed anything regarding the actual "shortcode"
Removed shortcode as it is currently not a very practical solution for webshops: They'd have to find a place in the html with a class or id where to attach the co2ok button. This is quite a burden on the customer(webshops), not very easy or quick.

Note: Since the intent was to make a shortcode option, the naming for this is currently still called shortcode, for clarity, this might need to be changed up.