Removed the reset of co2ok_checkout_placement AND co2ok_optout
Removed the reset of co2ok_checkout_placement AND co2ok_optout
There was a bug where if you post a save (any save) on the plugin settings page the co2ok checkout placement function would be reset to default.

NOTE: This was also the case for the co2ok optout function, so that was removed as well.  However this would usually be done on a different branch. It's done on this branch for efficiency.