Feature/shortcode option (#25)
* Added shortcode option in the Plugin settings and partially shortcode functionality

Everything works apart from the shortcode itself.
Also there appears to be a minor bug with the Choose Button Style part of the Plugin Settings page, I will have to check out whether this is due to the new feature or something with the browser cache

* The Button placement on Checkout page works

The user can now select between different area's on the Checkout-page where to place the button. This selection is done on the Plugin settings page.
It works - BUT - it's still buggy.
The placement doesn't seem to work properly everywhere (but in most places it does).
The button is placed over the "Place Order" button when selecting the review_order_after_submit placement, for example.
There is one issue where the button isn't shown anywhere (it might be that this hook is removed from Woocommerce altogether)

Also, currently when using the Placement selection and saving, it won't use the other options (Default, and Shortcode) anymore.
Warning: DON'T merge the current branch until this ^ bug is fixed.

Lastly, I've put in some debugging I will remove later.

* Finishing up the shortcode-option feature

Removing debugging
Temporarily turn off the shortcode option, but keeping the " Pick a Checkout Page location. " option.
Changed the description on the Plugin Settings page.
Removed an option which didn't show the button at all.
Temporarily turning off an option which appears visually buggy (the button overlapping the "Place Order" button on the checkout page)

Added action (hook) to also show the button on the Cart page with each checkout placement option.

* Removed anything regarding the actual "shortcode"

Removed shortcode as it is currently not a very practical solution for webshops: They'd have to find a place in the html with a class or id where to attach the co2ok button. This is quite a burden on the customer(webshops), not very easy or quick.

Note: Since the intent was to make a shortcode option, the naming for this is currently still called shortcode, for clarity, this might need to be changed up.

* renamed shortcode with placement

Due to the shortcode functionality being removed from the code, it's more adequate to change the naming from "co2ok_shortcode" to "co2ok_placement" for example.

* removed unused function

copy of renderCheckbox wasn't required. So I removed it.

* Changed description of the Button Placement

Changed the description of the Button Placement for the admin page (default.php)

* Removed everything related to co2ok_placement variable

Removed everything related to the co2ok_placement variable, only co2ok_checkout_placement variable is required for now.

* Removed the reset of co2ok_checkout_placement AND co2ok_optout

Removed the reset of co2ok_checkout_placement AND co2ok_optout
There was a bug where if you post a save (any save) on the plugin settings page the co2ok checkout placement function would be reset to default.

NOTE: This was also the case for the co2ok optout function, so that was removed as well.  However this would usually be done on a different branch. It's done on this branch for efficiency.