Feature/keyboard selectable (#24)
* Accessibility for screenreader and keyboard users

Added some a and button tags in template tags so you can "tab" through the co2ok button elements, it currently works entirely for the Default button, the minimal button will probably require some extra work in the CSS. (so this is currently still broken) - I've commented out the sections which I want to implement but aren't working as desired.
The chrome screen-reader can now read each relevant section for the Default button.

* Changed color of Make and compensation amount text

Changed color of make and compensation amount text.
However this feels out of place with the logo at the moment.
(Logo color also needs to change now)

Removed the outline around the button and selectable texts due to the <a> links to make it selectable with keyboard.

* Fixed keyboard selectable for the minimal button design

Now also works for minimal button design

* Reverted color back to gray for "Make" text and changed color of co2ok logo

"Make" text used to be gray, I turned it green so now I reverted it back to gray.
Changed the text of "ok" in the co2ok logo to suit the darkened color of the text. The logo is also turned slightly less blue and more green.