Removed buggy option for button Checkout Placement and visual bug for WP Flatsome Theme
After Order Notes option with button Checkout Placement placed extra instances of the button, and added extra instances when clicking the button. So the After order Notes is removed.

There was also a visual bug with the Flatsome WP Theme, this was solved by adding "none" to the wepkit appearance.