Gif/video feature works
Gif/Video box appears and video starts when the button is pressed and checked.
The video box disappears when the video ended.
It works for both Default and Minimal button design.
The only minor problem I can currently find is that the size of the box around the video on Mobile isn't correct.

Still needs to be done: Randomizer for several video's
Possibly a fade out for the videobox, rather than a quick removal.