mijnkraamshop css overwritting co2ok button
their store css has letter-spacing and and text-align that altered the layout of our button. Added two lines of css to counteract these theme settings

also, in chrome, their theme added spaces around + and € in compensataion amount. This caused issues with the button layout. To counteract this, a single line of javascript was added to remove spaces in compensation amount.